+++ DemoApp/Shared/StreamChatWrapper.swift
@@ -21,6 +21,7 @@ final class StreamChatWrapper {
         var config = ChatClientConfig(apiKeyString: apiKeyString)
         config.shouldShowShadowedMessages = true
         config.applicationGroupIdentifier = applicationGroupIdentifier
+        config.urlSessionConfiguration.httpAdditionalHeaders = ["Custom": "Example"]
         return config
     }()