+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [4.39.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.39.0)
+_October 05, 2023_
+
 ## StreamChat
 ### ✅ Added
 - Add support for custom HTTP headers in `ChatClientConfig.urlSessionConfiguration` [#2818](https://github.com/GetStream/stream-chat-swift/pull/2818)